+++ src/main/java/gs/mclo/fabric/commands/Command.java
@@ -1,0 +1,10 @@
+package gs.mclo.fabric.commands;
+
+import com.mojang.brigadier.builder.LiteralArgumentBuilder;
+import net.fabricmc.fabric.api.client.command.v2.FabricClientCommandSource;
+import net.minecraft.server.command.ServerCommandSource;
+
+public interface Command {
+    LiteralArgumentBuilder<FabricClientCommandSource> buildClient(LiteralArgumentBuilder<FabricClientCommandSource> builder);
+    LiteralArgumentBuilder<ServerCommandSource> buildServer(LiteralArgumentBuilder<ServerCommandSource> builder);
+}